Move form for creating a point into PointFormFields
